Fix numerous warnings and some spelling mistakes.

+++ src/CouchObjC/SBCouchServer.m
@@ -59,7 +59,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 - (NSString*)version
 {
-    NSString *server = [NSString stringWithFormat:@"http://%@:%u", self.host, self.port];
+    NSString *server = [NSString stringWithFormat:@"http://%@:%lu", self.host, (unsigned long)self.port];
     NSURL *url = [NSURL URLWithString:server];
     NSMutableURLRequest *request = [NSMutableURLRequest requestWithURL:url];
 
@@ -85,7 +85,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 - (BOOL)createDatabase:(NSString*)db
 {
     NSString *escaped = [db stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding];
-    NSString *server = [NSString stringWithFormat:@"http://%@:%u/%@", self.host, self.port, escaped];
+    NSString *server = [NSString stringWithFormat:@"http://%@:%lu/%@", self.host, (unsigned long)self.port, escaped];
     NSURL *url = [NSURL URLWithString:server];
 
     NSMutableURLRequest *request = [NSMutableURLRequest requestWithURL:url];
@@ -103,7 +103,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 - (BOOL)deleteDatabase:(NSString*)db
 {
     NSString *escaped = [db stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding];
-    NSString *server = [NSString stringWithFormat:@"http://%@:%u/%@", self.host, self.port, escaped];
+    NSString *server = [NSString stringWithFormat:@"http://%@:%lu/%@", self.host, (unsigned long)self.port, escaped];
     NSURL *url = [NSURL URLWithString:server];
     
     NSMutableURLRequest *request = [NSMutableURLRequest requestWithURL:url];
@@ -121,7 +121,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 - (NSArray*)listDatabases
 {
-    NSString *server = [NSString stringWithFormat:@"http://%@:%u/_all_dbs", self.host, self.port];
+    NSString *server = [NSString stringWithFormat:@"http://%@:%lu/_all_dbs", self.host, (unsigned long)self.port];
     NSURL *url = [NSURL URLWithString:server];    
     NSMutableURLRequest *request = [NSMutableURLRequest requestWithURL:url];
     
@@ -149,11 +149,11 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 -(NSString*)serverURLAsString
 {
-    return [NSString stringWithFormat:@"http://%@:%u", self.host, self.port];  
+    return [NSString stringWithFormat:@"http://%@:%lu", self.host, (unsigned long)self.port];
 }
 
 -(NSString*)name{
-    return [NSString stringWithFormat:@"%@:%u", self.host, self.port];  
+    return [NSString stringWithFormat:@"%@:%lu", self.host, (unsigned long)self.port];  
 }
 
 @end
